Create minimal sample for app build target
